Refactor moving Dcache to a new namespace
